Neovim cleanup + switching to Neomake
Neomake is doing what Syntastic was doing, but async.

I used a different pattern to configure the plugins so the
configurations (which sometimes need to run after initialization).

When I switched to Neomake it pointed out a bunch of lint issues in my
vimscript, so I fixed them too... Which is why this is a big commit.